refactor(agent): Centralize Anthropic 3-Tier Long Term Memory Mechanic
