+++ src/agents/builtin/memory/anthropic_tier.rs
@@ -6,23 +6,23 @@ use tokio::fs;
 /// 2) Detailed topic files (pulled on demand)
 /// 3) Raw transcripts (accessed via search only)
 #[derive(Clone, Debug)]
-pub struct Anthropic3TierMemory {
+pub struct Anthropic3TierMemoryStore {
     _base_dir: PathBuf,
     index_file: PathBuf,
     topics_dir: PathBuf,
     transcripts_dir: PathBuf,
 }
 
-impl Anthropic3TierMemory {
-    pub async fn new<P: AsRef<Path>>(base_dir: P) -> std::io::Result<Self> {
+impl Anthropic3TierMemoryStore {
+    pub fn new<P: AsRef<Path>>(base_dir: P) -> std::io::Result<Self> {
         let base_dir = base_dir.as_ref().to_path_buf();
         let index_file = base_dir.join("index.md");
         let topics_dir = base_dir.join("topics");
         let transcripts_dir = base_dir.join("transcripts");
 
-        fs::create_dir_all(&base_dir).await?;
-        fs::create_dir_all(&topics_dir).await?;
-        fs::create_dir_all(&transcripts_dir).await?;
+        std::fs::create_dir_all(&base_dir)?;
+        std::fs::create_dir_all(&topics_dir)?;
+        std::fs::create_dir_all(&transcripts_dir)?;
 
         Ok(Self {
             _base_dir: base_dir,
@@ -33,6 +33,21 @@ impl Anthropic3TierMemory {
     }
 
     /// Appends a new entry to the lightweight index.
+
+    pub async fn update_index(&self, content: &str) -> std::io::Result<()> {
+        tokio::fs::write(&self.index_file, content).await
+    }
+
+    pub async fn get_lightweight_index(&self) -> Result<String, String> {
+        if self.index_file.exists() {
+            tokio::fs::read_to_string(&self.index_file)
+                .await
+                .map_err(|e| e.to_string())
+        } else {
+            Ok(String::new())
+        }
+    }
+
     pub async fn append_to_index(&self, summary: &str, tags: &[String]) -> std::io::Result<()> {
         let char_count = summary.chars().count();
         let truncated_summary = if char_count > 150 {
@@ -125,6 +140,164 @@ impl Anthropic3TierMemory {
     }
 }
 
+
+
+#[async_trait::async_trait]
+impl crate::tools::anthropic_memory::MemoryAccessor for Anthropic3TierMemoryStore {
+    async fn write_topic(&self, topic_name: &str, content: &str) -> Result<(), String> {
+        let safe_name =
+            topic_name.replace(|c: char| !c.is_alphanumeric() && c != '_' && c != '-', "");
+        let path = self.topics_dir.join(format!("{}.md", safe_name));
+        tokio::fs::write(&path, content)
+            .await
+            .map_err(|e| e.to_string())?;
+
+        let mut existing_index = self.get_lightweight_index().await?;
+        let char_count = content.chars().count();
+        let truncated_content = if char_count > 150 {
+            let truncated: String = content.chars().take(147).collect();
+            format!("{}...", truncated)
+        } else {
+            content.to_string()
+        };
+        let new_entry = format!(
+            "- {}: {}\n",
+            safe_name,
+            truncated_content.replace(char::from(10), " ")
+        );
+        if !existing_index.contains(&safe_name) {
+            existing_index.push_str(&new_entry);
+            self.update_index(&existing_index).await.map_err(|e| e.to_string())?;
+        }
+        Ok(())
+    }
+
+    async fn retrieve_topic(&self, topic_name: &str) -> Result<String, String> {
+        crate::tools::anthropic_memory::MemoryAccessor::retrieve_topic(self, topic_name).await
+    }
+
+    async fn search_transcripts(&self, query: &str, limit: usize) -> Result<Vec<String>, String> {
+        crate::tools::anthropic_memory::MemoryAccessor::search_transcripts(self, query, limit).await
+    }
+}
+
+#[async_trait::async_trait]
+impl crate::memory_store::LongTermMemory for Anthropic3TierMemoryStore {
+    async fn store_session_message(
+        &self,
+        session_id: &str,
+        role: &str,
+        content: &str,
+    ) -> Result<(), String> {
+        let turn = format!("{}: {}", role, content);
+        self.append_transcript(session_id, &turn).await.map_err(|e| e.to_string())
+    }
+
+    async fn search_session_messages(
+        &self,
+        _session_id: &str,
+        query: &str,
+        limit: usize,
+        _summarize: bool,
+    ) -> Result<Vec<String>, String> {
+        crate::tools::anthropic_memory::MemoryAccessor::search_transcripts(self, query, limit).await
+    }
+
+    async fn retrieve(&self, query: &str, limit: usize) -> Result<Vec<String>, String> {
+        let mut results = Vec::new();
+
+        if !self.topics_dir.exists() {
+            return Ok(results);
+        }
+
+        let mut dir = tokio::fs::read_dir(&self.topics_dir)
+            .await
+            .map_err(|e| e.to_string())?;
+        while let Ok(Some(entry)) = dir.next_entry().await {
+            let content = tokio::fs::read_to_string(entry.path())
+                .await
+                .map_err(|e| e.to_string())?;
+            if content.to_lowercase().contains(&query.to_lowercase()) {
+                results.push(content);
+                if results.len() >= limit {
+                    break;
+                }
+            }
+        }
+
+        Ok(results)
+    }
+
+    async fn store(&self, content: &str, tags: Vec<String>) -> Result<(), String> {
+        let mut existing_index = self.get_lightweight_index().await?;
+
+        let char_count2 = content.chars().count();
+        let truncated_content = if char_count2 > 150 {
+            {
+                let truncated: String = content.chars().take(147).collect();
+                format!("{}...", truncated)
+            }
+        } else {
+            content.to_string()
+        };
+
+        let tags_str = if tags.is_empty() {
+            String::new()
+        } else {
+            format!(" [{}]", tags.join(", "))
+        };
+        let new_entry = format!("- {}{}\n", truncated_content.replace('\n', " "), tags_str);
+
+        existing_index.push_str(&new_entry);
+        self.update_index(&existing_index).await.map_err(|e| e.to_string())?;
+
+        Ok(())
+    }
+
+    async fn get_lightweight_index(&self) -> Result<String, String> {
+        Anthropic3TierMemoryStore::get_lightweight_index(self).await
+    }
+
+    async fn retrieve_topic(&self, topic_name: &str) -> Result<String, String> {
+        let safe_name =
+            topic_name.replace(|c: char| !c.is_alphanumeric() && c != '_' && c != '-', "");
+        let path = self.topics_dir.join(format!("{}.md", safe_name));
+        if path.exists() {
+            tokio::fs::read_to_string(&path)
+                .await
+                .map_err(|e| e.to_string())
+        } else {
+            Err(format!("Topic '{}' not found", safe_name))
+        }
+    }
+
+    async fn search_transcripts(&self, query: &str, limit: usize) -> Result<Vec<String>, String> {
+        let mut results = Vec::new();
+        let mut dir = tokio::fs::read_dir(&self.transcripts_dir)
+            .await
+            .map_err(|e| e.to_string())?;
+        while let Ok(Some(entry)) = dir.next_entry().await {
+            let content = tokio::fs::read_to_string(entry.path())
+                .await
+                .map_err(|e| e.to_string())?;
+            if content.to_lowercase().contains(&query.to_lowercase()) {
+                let filename = entry.file_name().to_string_lossy().to_string();
+                results.push(format!("Transcript {}:\n{}", filename, content));
+                if results.len() >= limit {
+                    break;
+                }
+            }
+        }
+        Ok(results)
+    }
+
+    fn as_anthropic_accessor(
+        &self,
+    ) -> Option<std::sync::Arc<dyn crate::tools::anthropic_memory::MemoryAccessor>> {
+        Some(std::sync::Arc::new(self.clone()))
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -133,7 +306,7 @@ mod tests {
     #[tokio::test]
     async fn test_lightweight_index() {
         let dir = tempdir().unwrap();
-        let memory = Anthropic3TierMemory::new(dir.path()).await.unwrap();
+        let memory = Anthropic3TierMemoryStore::new(dir.path()).unwrap();
 
         let summary = "This is a very long summary that should exceed one hundred and fifty characters so we can test the truncation logic properly. Let's add some more text just to be absolutely sure it goes over the limit.";
         memory
@@ -154,7 +327,7 @@ mod tests {
     #[tokio::test]
     async fn test_topics() {
         let dir = tempdir().unwrap();
-        let memory = Anthropic3TierMemory::new(dir.path()).await.unwrap();
+        let memory = Anthropic3TierMemoryStore::new(dir.path()).unwrap();
 
         // Write
         memory
@@ -174,7 +347,7 @@ mod tests {
     #[tokio::test]
     async fn test_transcripts() {
         let dir = tempdir().unwrap();
-        let memory = Anthropic3TierMemory::new(dir.path()).await.unwrap();
+        let memory = Anthropic3TierMemoryStore::new(dir.path()).unwrap();
 
         memory
             .append_transcript("session1", "User: Hello")
