+++ src/agents/builtin/memory_store.rs
@@ -998,52 +998,7 @@ mod tests {
         let _ = tokio::fs::remove_dir_all(dir).await;
     }
 
-    #[tokio::test]
-    async fn test_anthropic_3_tier_memory_store() {
-        let base_dir = "/tmp/test_anthropic_3_tier";
-        let _ = tokio::fs::remove_dir_all(base_dir).await;
 
-        let store = Anthropic3TierMemoryStore::new(base_dir).unwrap();
-
-        // Test lightweight index
-        store.update_index("Sample index content").await.unwrap();
-        let index = store.get_lightweight_index().await.unwrap();
-        assert_eq!(index, "Sample index content");
-
-        // Test topic retrieve
-        store
-            .write_topic("system_architecture", "Detailed DB schema")
-            .await
-            .unwrap();
-        let topic_content = store.retrieve_topic("system_architecture").await.unwrap();
-        assert_eq!(topic_content, "Detailed DB schema");
-        assert!(store.retrieve_topic("nonexistent").await.is_err());
-
-        // Test transcript search
-        store
-            .append_transcript(
-                "session1",
-                "User asked about memory.\n\nAgent replied 3-tier is better.",
-            )
-            .await
-            .unwrap();
-        store
-            .append_transcript(
-                "session2",
-                "User requested weather.\n\nAgent gave forecast.",
-            )
-            .await
-            .unwrap();
-
-        let res = store
-            .search_transcripts("3-tier is better", 10)
-            .await
-            .unwrap();
-        assert_eq!(res.len(), 1);
-        assert!(res[0].contains("Agent replied 3-tier is better."));
-
-        let _ = tokio::fs::remove_dir_all(base_dir).await;
-    }
 }
 
 #[async_trait]
@@ -1162,267 +1117,14 @@ impl LongTermMemory for PersistentMemoryStore {
     }
 }
 
-/// Anthropic 3-Tier Memory Store implementation. Mechanic: Anthropic 3-Tier Memory
-/// Crucial rule: Agent must treat memory as a "hint" and verify against actual state before acting.
-/// 1) Lightweight index (~150 chars/entry, always loaded in context)
-/// 2) Detailed topic files (pulled on demand)
-/// 3) Raw transcripts (accessed via search only)
-#[derive(Clone)]
-pub struct Anthropic3TierMemoryStore {
-    #[allow(dead_code)]
-    base_dir: std::path::PathBuf,
-    index_file: std::path::PathBuf,
-    topics_dir: std::path::PathBuf,
-    transcripts_dir: std::path::PathBuf,
+
 }
 
-impl std::fmt::Debug for Anthropic3TierMemoryStore {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Anthropic3TierMemoryStore").finish()
-    }
-}
 
-impl Anthropic3TierMemoryStore {
-    pub fn new<P: AsRef<std::path::Path>>(base_dir: P) -> Result<Self, String> {
-        let base_dir = base_dir.as_ref().to_path_buf();
-        let index_file = base_dir.join("index.md");
-        let topics_dir = base_dir.join("topics");
-        let transcripts_dir = base_dir.join("transcripts");
 
-        std::fs::create_dir_all(&base_dir).map_err(|e| e.to_string())?;
-        std::fs::create_dir_all(&topics_dir).map_err(|e| e.to_string())?;
-        std::fs::create_dir_all(&transcripts_dir).map_err(|e| e.to_string())?;
 
-        Ok(Self {
-            base_dir,
-            index_file,
-            topics_dir,
-            transcripts_dir,
-        })
-    }
 
-    pub async fn update_index(&self, content: &str) -> Result<(), String> {
-        tokio::fs::write(&self.index_file, content)
-            .await
-            .map_err(|e| e.to_string())
-    }
 
-    pub async fn write_topic(&self, topic_name: &str, content: &str) -> Result<(), String> {
-        let safe_name =
-            topic_name.replace(|c: char| !c.is_alphanumeric() && c != '_' && c != '-', "");
-        let path = self.topics_dir.join(format!("{}.md", safe_name));
-        tokio::fs::write(path, content)
-            .await
-            .map_err(|e| e.to_string())
-    }
-
-    pub async fn append_transcript(
-        &self,
-        session_id: &str,
-        turn_content: &str,
-    ) -> Result<(), String> {
-        let path = self.transcripts_dir.join(format!("{}.log", session_id));
-        use tokio::io::AsyncWriteExt;
-        let mut file = tokio::fs::OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(path)
-            .await
-            .map_err(|e| e.to_string())?;
-        file.write_all(format!("{}\n\n", turn_content).as_bytes())
-            .await
-            .map_err(|e| e.to_string())?;
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl crate::tools::anthropic_memory::MemoryAccessor for Anthropic3TierMemoryStore {
-    async fn write_topic(&self, topic_name: &str, content: &str) -> Result<(), String> {
-        let safe_name =
-            topic_name.replace(|c: char| !c.is_alphanumeric() && c != '_' && c != '-', "");
-        let path = self.topics_dir.join(format!("{}.md", safe_name));
-        tokio::fs::write(&path, content)
-            .await
-            .map_err(|e| e.to_string())?;
-
-        let mut existing_index = self.get_lightweight_index().await?;
-        let char_count = content.chars().count();
-        let truncated_content = if char_count > 150 {
-            let truncated: String = content.chars().take(147).collect();
-            format!("{}...", truncated)
-        } else {
-            content.to_string()
-        };
-        let new_entry = format!(
-            "- {}: {}\n",
-            safe_name,
-            truncated_content.replace(char::from(10), " ")
-        );
-        if !existing_index.contains(&safe_name) {
-            existing_index.push_str(&new_entry);
-            self.update_index(&existing_index).await?;
-        }
-        Ok(())
-    }
-
-    async fn retrieve_topic(&self, topic_name: &str) -> Result<String, String> {
-        let safe_name =
-            topic_name.replace(|c: char| !c.is_alphanumeric() && c != '_' && c != '-', "");
-        let path = self.topics_dir.join(format!("{}.md", safe_name));
-        if path.exists() {
-            tokio::fs::read_to_string(&path)
-                .await
-                .map_err(|e| e.to_string())
-        } else {
-            Err(format!("Topic '{}' not found", safe_name))
-        }
-    }
-
-    async fn search_transcripts(&self, query: &str, limit: usize) -> Result<Vec<String>, String> {
-        let mut results = Vec::new();
-        let mut dir = tokio::fs::read_dir(&self.transcripts_dir)
-            .await
-            .map_err(|e| e.to_string())?;
-        while let Ok(Some(entry)) = dir.next_entry().await {
-            let content = tokio::fs::read_to_string(entry.path())
-                .await
-                .map_err(|e| e.to_string())?;
-            for par in content.split("\n\n") {
-                if par.to_lowercase().contains(&query.to_lowercase()) {
-                    results.push(par.to_string());
-                    if results.len() >= limit {
-                        return Ok(results);
-                    }
-                }
-            }
-        }
-        Ok(results)
-    }
-}
-
-#[async_trait]
-impl LongTermMemory for Anthropic3TierMemoryStore {
-    async fn store_session_message(
-        &self,
-        session_id: &str,
-        role: &str,
-        content: &str,
-    ) -> Result<(), String> {
-        let turn = format!("{}: {}", role, content);
-        self.append_transcript(session_id, &turn).await
-    }
-
-    async fn search_session_messages(
-        &self,
-        _session_id: &str,
-        query: &str,
-        limit: usize,
-        _summarize: bool,
-    ) -> Result<Vec<String>, String> {
-        crate::tools::anthropic_memory::MemoryAccessor::search_transcripts(self, query, limit).await
-    }
-
-    async fn retrieve(&self, query: &str, limit: usize) -> Result<Vec<String>, String> {
-        let mut results = Vec::new();
-
-        if !self.topics_dir.exists() {
-            return Ok(results);
-        }
-
-        let mut dir = tokio::fs::read_dir(&self.topics_dir)
-            .await
-            .map_err(|e| e.to_string())?;
-        while let Ok(Some(entry)) = dir.next_entry().await {
-            let content = tokio::fs::read_to_string(entry.path())
-                .await
-                .map_err(|e| e.to_string())?;
-            if content.to_lowercase().contains(&query.to_lowercase()) {
-                results.push(content);
-                if results.len() >= limit {
-                    break;
-                }
-            }
-        }
-
-        Ok(results)
-    }
-
-    async fn store(&self, content: &str, tags: Vec<String>) -> Result<(), String> {
-        let mut existing_index = self.get_lightweight_index().await?;
-
-        let char_count2 = content.chars().count();
-        let truncated_content = if char_count2 > 150 {
-            {
-                let truncated: String = content.chars().take(147).collect();
-                format!("{}...", truncated)
-            }
-        } else {
-            content.to_string()
-        };
-
-        let tags_str = if tags.is_empty() {
-            String::new()
-        } else {
-            format!(" [{}]", tags.join(", "))
-        };
-        let new_entry = format!("- {}{}\n", truncated_content.replace('\n', " "), tags_str);
-
-        existing_index.push_str(&new_entry);
-        self.update_index(&existing_index).await?;
-
-        Ok(())
-    }
-
-    async fn get_lightweight_index(&self) -> Result<String, String> {
-        if self.index_file.exists() {
-            tokio::fs::read_to_string(&self.index_file)
-                .await
-                .map_err(|e| e.to_string())
-        } else {
-            Ok(String::new())
-        }
-    }
-
-    async fn retrieve_topic(&self, topic_name: &str) -> Result<String, String> {
-        let safe_name =
-            topic_name.replace(|c: char| !c.is_alphanumeric() && c != '_' && c != '-', "");
-        let path = self.topics_dir.join(format!("{}.md", safe_name));
-        if path.exists() {
-            tokio::fs::read_to_string(&path)
-                .await
-                .map_err(|e| e.to_string())
-        } else {
-            Err(format!("Topic '{}' not found", safe_name))
-        }
-    }
-
-    async fn search_transcripts(&self, query: &str, limit: usize) -> Result<Vec<String>, String> {
-        let mut results = Vec::new();
-        let mut dir = tokio::fs::read_dir(&self.transcripts_dir)
-            .await
-            .map_err(|e| e.to_string())?;
-        while let Ok(Some(entry)) = dir.next_entry().await {
-            let content = tokio::fs::read_to_string(entry.path())
-                .await
-                .map_err(|e| e.to_string())?;
-            for par in content.split("\n\n") {
-                if par.to_lowercase().contains(&query.to_lowercase()) {
-                    results.push(par.to_string());
-                    if results.len() >= limit {
-                        return Ok(results);
-                    }
-                }
-            }
-        }
-        Ok(results)
-    }
-    fn as_anthropic_accessor(
-        &self,
-    ) -> Option<std::sync::Arc<dyn crate::tools::anthropic_memory::MemoryAccessor>> {
-        Some(std::sync::Arc::new(self.clone()))
-    }
-}
 
 /// A simple implementation that stores memory in Redis using its list or sorted set capabilities.
 /// In a production system, this would likely use Redis Vector Search (RediSearch) or a dedicated vector DB.
@@ -2554,57 +2256,7 @@ mod get_conflicts_tests {
 mod anthropic_memory_tests {
     use super::*;
 
-    #[tokio::test]
-    async fn test_anthropic_3tier_memory_store_retrieve_and_store() {
-        let temp_dir = tempfile::tempdir().unwrap();
-        let store = Anthropic3TierMemoryStore::new(temp_dir.path()).unwrap();
 
-        // Initially index is empty
-        let index = store.get_lightweight_index().await.unwrap();
-        assert_eq!(index, "");
-
-        // Test storing multiple items
-        store
-            .store(
-                "User explicitly requested to use glassmorphism across all UI components.",
-                vec!["ui".to_string(), "design".to_string()],
-            )
-            .await
-            .unwrap();
-        store
-            .store(
-                "The PostgreSQL deployment requires enabling row-level security for multi-tenancy.",
-                vec![],
-            )
-            .await
-            .unwrap();
-
-        let index2 = store.get_lightweight_index().await.unwrap();
-        assert!(index2.contains("glassmorphism"));
-        assert!(index2.contains("[ui, design]"));
-        assert!(index2.contains("row-level security"));
-
-        // Add a mock topic file to test retrieve
-        store.write_topic("Database Architecture", "The database architecture relies heavily on PostgreSQL with row-level security enabled for multi-tenancy isolation. This is critical for data separation.").await.unwrap();
-        store
-            .write_topic(
-                "Frontend Style",
-                "Use flutter and glassmorphism. It should look modern.",
-            )
-            .await
-            .unwrap();
-
-        let results = store.retrieve("postgresql", 5).await.unwrap();
-        assert_eq!(results.len(), 1);
-        assert!(results[0].to_lowercase().contains("postgresql"));
-
-        let results2 = store.retrieve("glassmorphism", 5).await.unwrap();
-        assert_eq!(results2.len(), 1);
-        assert!(results2[0].to_lowercase().contains("flutter"));
-
-        let results3 = store.retrieve("nonexistent", 5).await.unwrap();
-        assert_eq!(results3.len(), 0);
-    }
 
     #[tokio::test]
     async fn test_cross_department_search() {
@@ -2836,99 +2488,7 @@ mod anthropic_memory_tests {
         let id: String = rows[0].try_get("id").unwrap();
         assert_eq!(id, "rec_override", "The correct record should remain");
     }
-    #[tokio::test]
-    async fn test_anthropic_3_tier_memory_flow() {
-        let dir = tempfile::tempdir().unwrap();
-        let store = Anthropic3TierMemoryStore::new(dir.path()).unwrap();
 
-        // Tier 1: Index
-        store
-            .store("User likes chocolate cake", vec!["preference".to_string()])
-            .await
-            .unwrap();
-        let index = store.get_lightweight_index().await.unwrap();
-        assert!(index.contains("User likes chocolate cake"));
-        assert!(index.contains("[preference]"));
-
-        // Tier 2: Topics
-        // Agent writes a topic
-        crate::tools::anthropic_memory::MemoryAccessor::write_topic(
-            &store,
-            "cake_preferences",
-            "User likes chocolate cake with strawberry frosting.",
-        )
-        .await
-        .unwrap();
-
-        // Agent retrieves the topic
-        let topic_content = crate::tools::anthropic_memory::MemoryAccessor::retrieve_topic(
-            &store,
-            "cake_preferences",
-        )
-        .await
-        .unwrap();
-        assert_eq!(
-            topic_content,
-            "User likes chocolate cake with strawberry frosting."
-        );
-
-        // Agent fails to retrieve non-existent topic
-        assert!(
-            crate::tools::anthropic_memory::MemoryAccessor::retrieve_topic(&store, "non_existent")
-                .await
-                .is_err()
-        );
-
-        // Tier 3: Transcripts
-        // Core loop stores session messages
-        store
-            .store_session_message("session_1", "user", "I would like to order a cake.")
-            .await
-            .unwrap();
-        store
-            .store_session_message("session_1", "agent", "Sure, what kind of cake?")
-            .await
-            .unwrap();
-        store
-            .store_session_message("session_1", "user", "Chocolate please!")
-            .await
-            .unwrap();
-
-        // Agent searches transcripts
-        let results = crate::tools::anthropic_memory::MemoryAccessor::search_transcripts(
-            &store,
-            "order a cake",
-            5,
-        )
-        .await
-        .unwrap();
-        assert_eq!(results.len(), 1);
-        assert!(results[0].contains("user: I would like to order a cake."));
-
-        let results_choc = crate::tools::anthropic_memory::MemoryAccessor::search_transcripts(
-            &store,
-            "Chocolate",
-            5,
-        )
-        .await
-        .unwrap();
-        assert_eq!(results_choc.len(), 1);
-        assert!(results_choc[0].contains("user: Chocolate please!"));
-
-        // Search should respect limit
-        store
-            .store_session_message("session_2", "user", "Chocolate is good.")
-            .await
-            .unwrap();
-        let results_limit = crate::tools::anthropic_memory::MemoryAccessor::search_transcripts(
-            &store,
-            "Chocolate",
-            1,
-        )
-        .await
-        .unwrap();
-        assert_eq!(results_limit.len(), 1);
-    }
 }
 
 #[cfg(test)]
